Implement q_insert_head function
The q_insert_head function was implemented to insert a new element at
the head of the queue. It ensures proper memory management by allocating
memory for both the new element and its string value.

The function uses strncpy to safely copy the string
and ensures null termination.

The new element is then added to the list using "list_add"
which has been written in list.h.

Change-Id: I6a3f6fab4983ab9865122008398e0837080d0852