Implement q_size
The q_size function was implemented to traverse the queue
and count the number of elements.
It iterates through the list using the list_for_each
macro from list.h and increments a counter for each element.

If the input head is NULL,
the function returns -1 to indicate an error.

Change-Id: I942c6f25573674c7a279cfdf29b87ca9970eb3c3